exposing GALOIS_ATTRIBUTE_ALIGN_CACHE_LINE as an alternative to CacheLineStorage<T>, which can be used to place alignment constraints on any variable.
More importantly, will force an object to be cache line aligned, if the 
alignment constraint is put on first member. Often an atomic/lock is used to 
synchronize/protect for modifications to other members of the class, and this lock
is desired to be on its own cache line. Enforcing alignment contraint on first member in such cases
will minimize the padding that needs to be added to the object.



git-svn-id: https://svn.ices.utexas.edu/repos/iss/projects/GaloisCpp/trunk@11814 522c16a8-e32f-0410-b4cc-8071c0b41197