Bump to C++17 standard
Recently there have been issues with aligned structures but without
aligned allocations. This can create segfaults as a compiler might
generate code assuming a certain alignment (e.g., vmovaps) but the
allocator does not preserve the same alignment [1].

This really seems to be an issue with ::operator new pre-C++17 and the
most stable fix is to move to a newer standard.

[1] https://stackoverflow.com/questions/52000359