Fix issue when CI doesn't run when a PR is created
The name of the state when a PR is first made is "opened" not "created".

When no states are specified, the action runs on "opened", "reopened"
and "synchronized":

  https://help.github.com/en/actions/reference/events-that-trigger-workflows#pull-request-event-pull_request

This does cause the CI to run twice when there is a push to a branch in
this repo and the branch also has a corresponding PR, but there doesn't
seem to be a straightforward way to avoid this while also supporting
fork-based development, which normally triggers PR opened and
synchronized events but not push events (on the upstream repo).