[libgalois] use the correct size for edges
I seem to have just moved the confusion (from the names of these
variables to myself) rather than eliminating it.

Didn't catch it in the quick validation that was done in patch
re-organization because the files I tested with are long enough to
accommodate the over sized reads and it still produces correct results
since we were technically fetching the data that was asked for. Worst
case was double the size of the edge array which is backed by the edge
data section.

It was a bummer for performance since the amount of data read by each
host increased linearly with their ID.

Signed-off-by: Tyler Hunt <thunt@katanagraph.com>