build: Remove app function
Applications can be written inside the source tree or outside of it. In
order to provide a consistent interface, we should have minimal
entanglement of the applications inside the source tree with the build
of the core library.

Inline usage of the app function with its definition to decouple
applications from the core library build system and each other.