Move PerThreadTimer to galois::runtime
Threads are a low-level primitive and the high-level galois namespace
should probably avoid frequent reference to it. Additionally
PerThreadTimer is (only) used by the runtime. We should minimize
dependencies from the galois::runtime to the galois namespace as ideally
the latter should be expressed in terms of the former rather than the
other way around.

Along the way, I moved as much implementation to the cpp file in an
effort to reduce include dependencies. Previously, Timer.h included both
ctime and chrono. Now it just includes chrono, the standard C++
definition of time.