tools: Be more tolerant of edge list inputs
Previously, graph-convert was strict about the format of edge list files
because there is no actual specification of this format, and it is
easier to support one variation rather than many.

However, most users of the tool have different expectations. Since there
is no documentation on what we think is a proper edge list file is it is
difficult to expect users to conform to our expected input either.

Instead, make graph-convert a little more lenient while also doing more
input validation. Now, it will check that lines contain pairs of triples
of numbers (depending on whether there is edge data), and it will skip
any line that doesn't.

This conveniently handles various commenting conventions in edge list
files.

Along the way, fix a bug where a duplicate edge is added when the input
contains a blank line.