Remove ConditionalReduction.h
Most use cases of conditional reduction seem to be better characterized
as optional application-specific profiling, for which conditional
reduction is not exactly the desired abstraction, e.g., profiling
benefits from having uniform output (for further data analysis) and
common enable/disable configuration.

Until these use cases are mapped out and a general interface determined,
it is better to keep these concepts out of the core library and leave it
for applications to implement these ideas on their own.