Die when num edges exceeds limit
If the number of edges for a node exceeded the maximum number of edges
that could fit on a page, LC_Morph_Graph would silently allow the buffer
overrun. Change this behavior to explicitly die instead.

Also, previously, local_edges->begin[k] != local_edges->end for any k
because the alignment of local_edges->end did not account for the
alignment of local_edges->begin.