Align PerThreadStorage memory
PerThreadStorage did not account for alignment of its element types.
Since PerThreadStorage is almost always used for types that have cache
line alignment, adopt the simple fix of rounding up any allocation to
the cache line size.

I considered an implementation that dealt with alignment explicitly but
it would complicate the allocation and deallocation logic which would
only matter if the PerThreadStorage arena were nearing capacity, which
is currently uncommon. It seems reasonable to defer more complicated
bookkeeping strategies until there is a stronger reason to improve this
functionality.