fixed bufferwrapper vector access
Originally the raw_mem pointer is initialized to the same data as the
dummy vector. However, the dummy vector's data pointer is updated at
some point, causes the old initiailization to be useless. Changed
behvaior such that if dummy vector is initialized, then access is always
from it. TODO how much overhead is the check?