Layer phases + graph accessors
Added more accessors to the graph like node iterators and LID-GID
functions. More importantly, added the concept of a layer phase (test,
validate, train) which causes training/masking to occur differently.
Added a function to the graph object to check if an LID is part of a
particular group of nodes to train as well.