GNN softmax loss output layer
This commit adds an implementation of the Softmax Loss output layer. The
layer creates a probability distribution of the rows of a passed in
matrix. The backward phase returns gradients on how to best move
computation towards a distribution that favors the ground truth.

Also includes some minor cleanup of the GCN layer.