Softmax layer test
Adds a test that runs the forward and backward phase of the softmax
layer. Verification only included  for the forward pass at this point:
it checks to make sure the probability distribution is as expected from
the test input. The backward phase is just run without checking its
output for sanity purposes.