Add accessors and type to GNNLayers
Adds a few more accessors to GNNLayer classes as well as a type field
which can be used to determine what kind of layer an object is. Note the
separation between an output layer and an intermediate layer; this may
be merged later into a single field, but the current design is that an
intermediate and output layer are considered differently throughout the
codebase.