Graph neural network construction + unit test
Adds initial implementation of a graph neural network's constructor
which can construct an arbitrary GNN given some config object and a few
other details. Adds a unit test which checks the structure of the GNN to
make sure that it is sane.

Also adds a feature length accessor to GNNGraph.

Next step is the implementation of forward/backward phases in the GNN.