GNN backward phase; no weight update
Added implementation for GNN backward phase without the weight update.
Each layer passes its gradients to the previous layer in the GNN where
weight gradients as well as gradients for the next layer in propagation
are determined.

The weight gradients are not yet used to update the weights via SGD.
This requires the addition of optimizer classes and will be the next
step in this refactoring.

Adds a call to the backward phase in the GNN FB test to make sure no
crashes occur.