Added gradient descent to backward pass
GNN's backward pass now also calls the optimizer to update weights based
on the gradient, thus completing the pipeline for a full training epoch.
The tests have been updated accordingly to fix the adam size (supposed
to be size of layer, not output).

Next step is to add accuracy measures and then a full program can be
tested end to end.