Global accuracy in GNN; test for it as well
Added a function to the GNN class to get accuracy of a prediction
distribution from a softmax layer based on training, validiation, and
test modes in the GNN. Added a field to track the mode in the GNN in
addition to it being tracked in the layers.

Added a test to make sure accuracy readings were returning correctly as
expected as well.