epoch-test
Test that runs on the cora dataset for 100 epochs to see if accuracy
grows as time passes.

The test has exposed some issues with the dropout option that will need
to be fixed in a future commit, but otherwise training accuracy seems to
grow as time passes and the program seems to scale.