Option to disable agg/update flip, reenable tests
Adds an option to layers to make it so program doesn't automatically
flip aggregate/update to avoid more work in the aggregate step. This
allows the tests that broke before to be renabled. Also fixed gnnfb
test by removing back prop call in val and test (you should not call
those during those phases; back is only for training).