GNN weight gradient synchronization + unit test
Adds a wrapper over any vector to make it possible to sync the vector
using Gluon: assumes the vector is replicated present on all hosts.

Use this wrapper to synchronize weight gradients calculated on each host
during the backward phase of computation. This commit includes a sum and
an average sync function with the average one being used by default.
TODO later is to make this user configurable.

Adds a unit test to make sure the weight gradient works as expected.

With this commit an end to end training framework should work. There are
still some accuracy issues to resolve, however (not matching old code
accuracy).