Command line options for GNN apps (libgnnbench)
Added a new static library in the lonestar directory called libgnnbench
that will be the backend for the (distributed) gnn benchmarks added in
the future. At the moment it contains command line declarations for user
configurable things in the GNN.

Also changed the variable name of the variable containing the default
path for gnn inputs (preparation for letting it get passed in as a user
defined parameter).