Start code for GNNBench, new dist gcn app
Adds the commandline parsing/stats setup code to GNNBench (based on
DistBench). Adds a new gcn-dist app under lonestart/gnn/distributed that
will be the main executable for gcn from now on. Successfully compiles
using the GNNBench code.

Next step is the GNN construction code in the lonestar end of the code.