Disable GALOIS_SUPPORT_ASYNC in libdist
The SUPPORT_ASYNC flag makes messages use issend instead of isend.
issend was causing issues with distributed execution where under
certain interleavings of send/recv a send would get presumably
corrupted. For now, disable issend until a fix is found (or maybe
remove it forever depending on if it's really necessary or not).

Performance wise it does not seem to show much of an impact on GNN
apps; mostly noise and large variance has been observed.