Fix GNNGraph default directory init bug
The constructor for GNN Graph that does not take an input directory was
constructing a GNNGraph in its function scope and not constructing the
actual caller of the constructor. This led to most of the unit tests
failing as the graph was not constructed correctly.

Moral: run unit tests before every commit.