Allocate features of graph on GPU (first test)
Adds initial CUDA headers/impl to GNNGraph, starting with a struct that
is meant to hold all GPU based allocations for the graph. This is
compiled separately from the main GNN library using nvcc and linked into
the main GNN library.

This commit adds an Init GPU call that is meant to initialize all GPU
memory used for GNN training. For now, it only allocates/copies over the
features of the graph. The rest are incoming in later commits.

Also contains a change to the Gluon gradient interface which adds a few
dummy typedefs so that Gluon doesn't complain (GPU build expects certain
typedefs to exist due to the marshal graph function).