Copy over int graph topology onto GPU for GNN
Adds code to cast the CSR of a partitioned graph into ints and copy it
over to the GPU.

The use of ints is because BLAS CSR standard is apparently ints, so this
is a future-proofing step for using GPU BLAS calls.

Also adds the destructor for GPU graph allocations which frees up
allocated memory + saves more variables onto GPU memory like feature
length, node count, etc.