Massive signature overhaul for vectors in GNNs
Previously in most of the GNN code std::vectors were being passed around
for the main calls to training/inferencing. The problem with this is
that GPUs do not have std::vector, but only pointers. To avoid losing
out on the size call, a new class called PointerWithSize has been added
that wraps a pointer and attaches a size to it that can be accessed via
size. It also has a similar interface to a std::vector so as little of
the code has to be changed as possible.

This class is now used instead of std::vector in the main
Forward/Backward Phase calls in the GNN in preparation for the GPU
support. This also required a few related changes such as making some
functions non-const (because the PointerWithSize is not a const
construction since it grabs the raw pointer).

Most of the tests have been edited as well to account for the new return
type.