Disable majority GNN tests for GPU build
Turn off a majority of the GNN tests for the GPU build because the data
is on the GPU so doing a check of the data would involve copying it back
to the CPU: this will be done later.

This commit is unlikely to build as it does not include a few changes
coming in later commits that make things work.