Use PointerWithSize for functions in layers
Since the code now needs to run with both CPUs and GPUs, the
"orchestration" code cannot use vectors anymore. This commit replaces
them with PointerWithSize which can be used by both CPUs and GPUs. This
commit adds initialization code for these objects depending on the
build.

This commit also begins to split the calls in CPU and GPU variants: the
orchestration code remains the same, but depending on the build, the
code will either call a CPU version or a GPU version. This split is done
for dropout and aggregation so far (though this commit does not
include the implementation of the GPU code; just the call).