GCN aggregation code on GPU
This commit adds the GCN aggregation code on the GPU: each warp gets a
node, and the warp splits the summation of its feature among its
threads. Therefore, the implementation works well if the feature vector
is big to take advantage of the paralellism.

This code was taken from the old GPU code, so it may not necessarily be
the best optimized just yet. cuSparse may perform better as well, but at
the moment I just want to get something working.