Adds CuBLAS SGEMM function
This commit adds a function to call CuBLAS's SGEMM for doing matrix
multiplies on the GPU using row-major matrices. Note the way matrices
are passed into the function: B * A instead of A * B. The idea is that
the CuBLAS function assumes column-major, so using row-major matrices
makes it so that it's a transpose matrix. (BA)^T = C^T, and C^T in
column-major form is a row-major C (exactly what we want).