UpdateEmdeddingsGPU function added
This commit splits the update emdeddings call in the GCN layer into a
CPU and GPU version. The GPU version calls a function which at the
moment will call into cuBLAS to do the multiplication.

Next step is to unit test the forward pass to make sure the results are
sane.

This commit also includes a clang-format run on SoftmaxLayer (which
currently is not being used anywhere anyways as it was an example Xuhao
added).