GPU conv layer forward pass test
Reenable assertions for the GPU GCN layer forward pass which now works.
The next step is to get the backward pass test working which involves
adding a function for user code to copy things over to CUDA without
needing to include the CUDA header + adding the appropriate GPU
functions in the backend.