Definition of forward phase gpu softmax
This commit adds the softmax/cross entropy function to GNNMath.cu and
uses it to define the GPU Softmax forward phase function. An additional
argument was added to the forward phase gpu call to deal with the
different phases: the phase argument details which mask to use in the
softmax. There are a few things left to do that will be done later,
namely zero'ing out the output matrix.

Note that I have NOT defined cross entropy for the forward phase: it is
only used to calculate loss, and I'm not using loss nor referring to it
anywhere in my code or analysis at the moment..