Softmax forward phase fixes; unit test for it too
Fixed some bugs exposed by the unit test for softmax forward, namely
that the feature length size was incorrect and that the vector was not
being 0'd out before softmax occured. The unit test in question has been
ported over from the cpu softmax unit test as well.

The next step is to finish up the backward pass for the softmax layer
and reactivate the unit test calls to the backward phase. I also need to
consider actually checking backward phase output to make sure it is
sane.