Fixed ground truth type on GPUs
Ground truth is represented with GNNLabel, but I was using a GNNFloat.
This caused the labels being read to be garbaged when used on the GPU.
This commit changes it them to the correct type.

It also includes the signature definition of the backward phase: the
implementation will be included in the next commit. (Split the commits
up for modularity's sake)