Softmax backward phase on GPU
Adds the backward phase for the softmax layer for the GPU. The
implementation is taken from the non-refactored old code: it copies a
prediction to shared memory (presumably to improve locality) then does
cross entropy to softmax derivatives. It remains to be seen if the
shared memory copy is actually more efficient; some testing will be done
down the line.

Also adds print to both cpu and gpu softmax tests in order to verify
that both are doing the same compute (which they are in this commit).