Structure added for GPU global accuracy function
This commit adds the declarations for the global accuracy
getter for GPU GNNs as well as the orchestration of the call to the GPU
version. The rest of the implementation will come in a later commit: for now
this isn't priority as I can still compute accuracy on the CPU.

Adds a new GNNGPU object to hold all GPU related things for the GNN
class.