Allocate GPU memory for Adam optimizer
Adds a GPU Adam optimizer class that holds the allocations for the
moments used in the adam optimizer on the GPU. Adds a gpu version of the
adam test as well to make sure build is sane in its current state.
The CPU optimizer class is also now split into the CPU/GPU paths
depending on which build is being used.

Next step is to do the adam optimizer on the GPU proper.