gpu epoch test; fixes to returned pointers
Adds a gpu version of the epoch test and fixes the pointers returned
from a GNN layer (it was always returning CPU pointers even in the GPU
build). Adds error checking to cuSparse call too.

gpu-epoch-test runs a GNN end to end (still missing some features that
CPU has), but it has to copy predictions over from GPU (slow, should do
this from GPU end) + there seem to be accuracy issues on reddit. Will be
resolved in a later commit.