GNNLayer GPU: init dropout memory
Initializes a dropout mask for every GPU layer. Can be optimized if
dropout is disabled (i.e. do not allocate) for both CPU/GPUs. This will
be handled later once a base implementation of everything is settled.

It is a float because the float will be checked during dropout to see if
it crosses some threshold for dropout.