GNN GPU: forward phase dropout
Adds code necessary for forward phase of dropout. Adds a data structure
for storing chars (result of dropout rng) and renames vars accordingly
(floats are converted to bools depending on threshold). Fixes the
variables passed into the dropout call as well as before it was using
CPU pointers: this commit changes the pointer to use the PointerWithSize
objects allocated for this purpose.