GPU GNN: backward phase dropout derivative
Adds the GPU wrapper + kernel call for the derivative of dropout for the
backward phase of the GCN layer. Makes the epoch tests for both CPUs and
GPUs equivalent as well.

With this commit, single CPU/GPU functionality for the GCN is roughly
the same. Note that answers will not be the same due to dropout RNG
being different on the 2 platforms.