Gradient sync to sum, master only softmax
Changes the gradient sync function to use sum instead of average, and
make it so the softmax layer only loops over the master nodes on each
host.

The effects of this are that accuracy in a distributed setting will be
exactly the same as accuracy in a single host setting because no
redundant computation will occur. In practice, however, RNG on each host
(i.e., dropout) will cause distributed execution to differ from single
host execution. Turning off all RNG will make it so the exact same
computation occurs (tradeoff is that dropout isn't done, so overfitting
to the train set may occur).