Partial SAGE layer + self loop norm fix in GCN
Adds incomplete SAGE layer implementation: the mean aggregation (done
via the norms being 1 / degree) is done in the SAGE layer. What needs to
be done next is the concat part of it. (next commit will take care of
this)

Also fixes a minor thing in GCN self-loop aggregation where norm^2 needs
to be applied to the self feature.