train/val/test splits for non-complete masks
Code before did not support masks that were not complete (i.e., assumes
that range given in file is complete meaning if in range it is part of
that set). Code now checks to make sure this is the case before doing
that; if not, it uses the mask instead (slower, but correct).