SAGE Layer completed + tests
This commit finishes up the implementation of the SAGE layer's backward
phase. One thing to note is that if dropout is enabled, then the dropout
matrix needs to be used during backward phase computations and not the
original. This needs to be fixed in the other existing layers.

GNNMath has a new version of the SGEMM routine that aggregates directly
into the output matrix: this is useful for the SAGE split matrix.