Fixed aggregate sync in GNNs
The buffer wrapper I was using to save a copy of data during sync is
incorrect as the memory can potentially get written before it is
actually serialized into a message. This was leading to inconsistent and
wrong results for GNN training. This commit changes it to use a gstl
vector. There is now some copy overhead due to this, but the tradeoff is
actual correct execution.

Some small fixes to the aggregate sync test as well.