Softmax-cross-entropy derivative merged
Merged the application of softmax and cross entropy derivatives into a
very simple subtraction term that is apparently correct according to
math gurus on the internet.

Advantages:
1) much simpler
2) most importantly, numerical stability + correctness: previous
softmax/cross-entropy derivative was causing blowup of accuracy after a
few epochs where Xuhao OMP code was not. This simplifies the derivative
into something that is numerically stable (and probably more correct
than what was there before).