better serialization for vec of vectors in gluon
Vector of vector serializatoin in Gluon was bad. This commit fixes it by
serializing all vectors into a single one before sendoff so
serialization unit doens't have to deal with it.

Because this commit was made in a rush (so i can schedule tests) I did
not really bother cleaning it much.