The big GNN space saving commit (part 1)
This commit does the following:

- Temporary/memoization matrices have been reduced to a minimum of only
1 per layer. The size of it will be the smaller of the input or output
matrices. This change along with distributed execution/masking required
some reorganization of the order of things in order to make it so you
can mask the correct things at the right time + avoid overwriting
something before it is used. This has made the code more annoying to
read as a result. Some cleanup will be necessary.

- Forward output matrices of layer n now act as the backward output
matrix for layer n + 1. This is a huge space savings, but again, it
required some changes to keep correctness. For example, for the
activation derivative I now have to keep a bitset marking which elements
were originally greater than one (originally this was done by checking
the forward output matrix itself, but now it gets overwritten so you
can't do that anymore).
This change required a signature change to the layer constructors.

- Tests now have to pass in a "forward output" for a layer to write the
backward output into. This has made things slightly more annoying to
read as well.

Part 2 will make it so output layers do not allocate their own output
matrices and just overwrite the input matrix. This will require some
function signature changes, which is why it will be done in a separate
commit.

Also TODO: Distributed SAGE unit tests.