GNN space saving (part 2): no softmax output
The softmax layer now reuses the output layer from the layer before it
in all of its operations, making it possible to completely avoid
allocating another matrix (resulting in *huge* space savings if # of
input classes is sufficiently large). The problem with this is now you
cannot check the output matrix of the layer before it (because softmax
will destroy it), and running backward phase will destroy the
predictions (meaning you have to check accuracy before you do gradient
descent).

Some tests had to be changed as a result of the changes described above.