GNNGraph now has edge sample bit and reverse edges
After graph partitioning, reverse edges get constructed on each
partition: this is to do correct aggregation later when sampling edges.
A byte has been added to all edges as well to mark if the edge is "on"
or "off" (i.e., sampled), with the corresponding in-edge sharing this
data as necessary.

This is all in preparation for distributed edge sampling.