Distributed calculation of gnn node degrees
Before this commit, degrees were calculated on every host by loading in
full original graph and using it. That is not efficient: this commit
makes every host count degrees and send it to all other hosts.
The reading of the whole graph topology has been removed.

Removes a bunch of older unused functions like SpecialNormFactor and
such as well.