GNN single host training minibatching
Allows for training using minibatching on a single host via a command
line argument.

A minibatcher class has been added that takes in a mask and batches
it via linear scan over it (O(n) ultimately per epoch). This mask
is then used to mark the seed nodes for the sampler (which can do full
sampling or it can take all nodes; this needs to be implemented).

The rest of execution works the same as previous sampling subgraphs:
main difference is the seed nodes aren't all training nodes, and
everything works out the same way from there.

Other notable changes
- GNNMask class for masks
- kBatch phase for minibatching; set the phase and entire pipeline
works as you might expect
- Removed old node sampling stuff because unused right now;
might bring back SAINT sampling if it turns out to be useful
- Some signature changes to functions