Minibatch generator loops through masters only
Minibatch generator should only loop through masters to avoid duplicates
in minibatches in a distributed setting. This assumes like all other
parts of the code at the moment are that master nodes always take the
prefix of node ids.

Generator takes the right bound of it and uses it instead of the mask
size.