WIP: optimize row size as layers progress
1) Seed choosing for sampling needs to sync the seeds across
hosts if only master nodes get sampled (which is what it has
been changed to if I remember correctly). Needs sync esp. if
minibatching since not all hosts know what other hosts are minibatching.

2) In minibatch setting, rows will decrease as some nodes will not
need to aggregate/not need xform. This has been implemented in
minibatching setting (but not others, since WIP).

3) To support #2 above, subgraph construction has changed so that
the SIDs are suffixed in the order that they are deactivated as
the rows progress. This increases construction time (how much is
something I need to check).

4) SAGE layer updated to deal with #2; input/output rows used depending
on phase.

NOTE: Same as other commits: SAGE only; GCN is super outdated at this
point.