GNN subgraph view implementation
Adds another form of subgraph creation: views. Does not construct the
explicit subgraph but only creates the LID-SID mappings. Issue is that
this moves the overhead to the forward/backward aggregation where the
program must now loop over a very large number of nodes + do mappings
from LID to SID in the background; per minibatch this overhead adds up.