GNN subgraphs always use global norm factor
Subgraph degree norm factor was buggy because I wasn't finding the
incoming degrees and using them for the subgraph; this caused accuracy
to get really weird depending on code route taken. For simplicity (and
to speedup subgraph construction) only global degrees are used even for
subgraphs.

This will 100% affect time to accuracy due to possible overcompensation
during training, but it saves a lot of micromanagement.