Back to using subgraph degrees
Subgraph degrees used again because accuracy when sampling suffers
otherwise. Don't actually need in-degrees because out-degree was used in
forward (meaning it's used in backward). Also, norm factor should
**NEVER** be 0 after some more thought. Added an assertion checking for
it in the debug build.

The other fix this commit includes is that the "choose all" mode must be
set appropriately when testing occurs (since test will always use all
degrees).