GNN: Optimize SID map, fix subgraph master masking
1) SID mapping made many passes over the graph; this is inefficient.
Optimize by counting how many nodes will make first appearance in each
layer and make only a single pass over the graph to get SID mappings.
Further improvement could be obtained by making this run in parallel
(doable with another prefix sum on top of existing one for work assigned
to each thread possibly), but will go there as necessary.

2) Master masking in SAGE/GCN layers for correctness was buggy in
subgraph case as it is possible that masters would appear beyond the
layer 0 prefix used by subgraphs. Avoid this by using a bitset which
indicates masters that are not in this prefix so that they are not
masked out accidentally. This fix seemingly improves time to accuracy as
well as time in general because less 0s are being written.