Sample sync; non subgraph node needs to return vec
Before this commit, sample sync for a node not in the subgraph would
return an empty vector. This is bad because Gluon will expect vectors of
a certain length, and trying to copy beyond allocated memory may cause
memory corruption. This fixes that by allocating dead space to serialize
anyways.