Sample flag sync can be readSource
Sample flag setting is readAny before this commit: rationale I had back
then was that even if a node didn't have outgoing edges it should still
be included in the graph. This actually isn't necessary though because I
remove dead mirrors later in subgraph construction, and it turns out
that this sync call is actually a major part of subgraph construction.

Therefore, to optimize runtime, go to readAny. This can cause problems
if some vertex is activated later without the sync (i.e. no consistent
"time" in which a vertex is entered), so in the sync call you have to
make sure not to set anything that exceeds the current layer's number of
rows (or you get a segfault).