Minibatch shuffling between epochs
First half of ogbn paper fix: minibatches need to be shuffled
between epochs to prevent overfitting to the same node minibatches
in every epoch.

The second half is input related: my remapping of the graph seems
to have broken something as using the original graph results
in much better accuracy. For the time being I will use that one.