GNN softmax, sync, masking correctness changes
1) You cannot use readSource for sampled nodes: they need to be active
across all hosts the moment that they become active on one host because
all other hosts need to start calculating the gradients for them in the
backward phase because they may be used later for gradient computation.
(unless they're *completely* disconnected in a subgraph, but that
is unknown until sampling is done; therefore, readAny must be used)

2) Softmax was using a "is valid for phase" check for nodes: this
is wrong because hosts will need to compute the softmax value
for nodes that aren't part of its own batch but part of another
hosts' batch.

3) A matrix was being masked incorrectly by SAGE due to a faulty
if condition: this broke the distributed execution accuracy curve.
The if has been fixed.

Performance wise, (1) and (2) will affect performance due to increased
sync volume and more compute in the softmax layer.