Major mirror deletion fix: don't delete masters
Master proxies were being deleted from subgraphs if they did not have
incoming/outgoing edges. This is a problem because the master is
responsible for reducing updates from all other proxies which may still
exist on other hosts, and this was the cause of accuracy degradation at
a higher number of hosts.

This problem did not appear much for edge cuts because master nodes
would not get deleted since edges end up on master.