WMD CSV-based graph ingestion in Galois (#3)
This commit is to enable Galois/Gluon to read WMD CSV-based graphs. This is a temporary code and this reads the whole graph on each machine on memory, which is not scaled. This will be updated to scalable graph ingestion with new dynamic graph data types (e.g., Log-structured CSR). For now, users can test WFs and ISBs with the WMD inputs through this.